Add initial R vs. Julia benchmarks
This commit adds a very simple R vs. Julia benchmark: simulating the
distribution of the sample mean of an arbitrary number of uniform
random variables. The R code is as idiomatic as I can make it, but
there might be better-performing options. The Julia code runs almost
4 times faster and is pretty straightforward, with only one "trick"
that accounts for a small part of the speed difference.